+++ IOptimization.pde
@@ -1,5 +1,5 @@
 public static interface IOptimization{
-  void optimize(IFunction function, float... point);
+  void optimize(IFunction function, double... point);
   String getInfo();
   void adjust(char command);
 }
