+++ MatyasFunction.pde
@@ -1,24 +1,24 @@
 public static class Matyas implements IFunction{
 
-  public float calculate(float... point){
-    float x = point[0];
-    float z = point[1];
+  public double calculate(double... point){
+    double x = point[0];
+    double z = point[1];
     
     return 0.26*(x*x+z*z)-0.48*x*z;
   }
   
-  public float[] gradient(float... point){
-    float x = point[0];
-    float z = point[1];
+  public double[] gradient(double... point){
+    double x = point[0];
+    double z = point[1];
     
-    return new float[]{0.26*2*x-0.48*z,0.26*2*z-0.48*x};
+    return new double[]{0.26*2*x-0.48*z,0.26*2*z-0.48*x};
   }
     
-  public float functionMax(){
+  public double functionMax(){
     return 10;
   }
   
-  public float functionMin(){
+  public double functionMin(){
     return -1;
   }
